wip u32

+++ packages/wgsl/src/geometry/tick.wgsl
@@ -1,8 +1,8 @@
-@export fn getTickSegment(index: i32) -> i32 {
-  let n = TICK_DETAIL + 1;
+@export fn getTickSegment(index: u32) -> i32 {
+  let n = u32(TICK_DETAIL + 1);
   let i = index % n;
-  if (i == 0) { return 1; }
-  if (i == n - 1) { return 2; }
+  if (i == 0u) { return 1; }
+  if (i == n - 1u) { return 2; }
   return 3;
 };
