+++ packages/wgsl/src/instance/draw/virtual-solid.wgsl
@@ -1,7 +1,7 @@
 use '@use-gpu/wgsl/use/types'::{ SolidVertex };
 use "@use-gpu/wgsl/use/color"::{ toColorSpace };
 
-@external fn getVertex(v: i32, i: i32) -> SolidVertex {};
+@external fn getVertex(v: u32, i: u32) -> SolidVertex {};
 
 struct VertexOutput {
   @builtin(position) position: vec4<f32>,
@@ -14,7 +14,7 @@ fn main(
   @builtin(vertex_index) vertexIndex: u32,
   @builtin(instance_index) instanceIndex: u32,
 ) -> VertexOutput {
-  var v = getVertex(i32(vertexIndex), i32(instanceIndex));
+  var v = getVertex(vertexIndex, instanceIndex);
 
   return VertexOutput(
     v.position,
